pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "astro-provider-ray"
authors = [
    { name = "Venkata Jagannath", email = "venkat@astronomer.io" },
    { name = "Astronomer", email = "humans@astronomer.io" },

]
license = { text = "Apache License 2.0" }
description = "An Apache Airflow provider package built by Astronomer to integrate with Ray."
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = [
    "version"
]
requires-python = "~=3.8"
dependencies = [
    "apache-airflow>=2.9",
    "ray[default]",
    "kubernetes",
    "requests",
    "pytest",
    "PyYAML",
    "pytest-asyncio"
]

[project.urls]
Homepage = "https://astronomer.io"
Source = "https://github.com/astronomer/astro-provider-ray/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "ray_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "ray_provider.__version__" }

[tool.setuptools.packages.find]
exclude = [
    "*tests.*",
    "*tests"
]

######################################
# THIRD PARTY TOOLS
######################################
[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310']

[tool.isort]
profile = "black"
known_third_party = ["airflow", "jinja2"]

[tool.mypy]
strict = true
ignore_missing_imports = true
no_warn_unused_ignores = true

[tool.ruff]
line-length = 120
[tool.ruff.lint]
select = ["C901", "I", "F"]
ignore = ["F541"]
[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.distutils.bdist_wheel]
universal = true
